+++ libicmp/ping.h
@@ -29,8 +29,8 @@ struct ping_data
 {
   int    ping_fd;        /* Raw socket descriptor */
   int    ping_type;      /* Type of packets to send */
-  int    ping_count;     /* Number of packets to send */
-  int    ping_interval;  /* Number of seconds to wait between sending pkts */
+  size_t ping_count;     /* Number of packets to send */
+  size_t ping_interval;  /* Number of seconds to wait between sending pkts */
   struct sockaddr_in ping_dest; /* whom to ping */
   char   *ping_hostname;     /* Printable hostname */
   size_t ping_datalen;   /* Length of data */
@@ -68,10 +68,10 @@ struct ping_data
 
 PING *ping_init (int type, int ident);
 void ping_set_type (PING *p, int type);
-void ping_set_count (PING *ping, int count);
+void ping_set_count (PING *ping, size_t count);
 void ping_set_sockopt (PING *ping, int opt, void *val, int valsize);
-void ping_set_interval (PING *ping, int interval);
-void ping_set_packetsize (PING *ping, int size);
+void ping_set_interval (PING *ping, size_t interval);
+void ping_set_packetsize (PING *ping, size_t size);
 int ping_set_dest (PING *ping, char *host);
 int ping_set_pattern (PING *p, int len, u_char *pat);
 void ping_set_event_handler (PING *ping, ping_efp fp, void *closure);
